Move buffer::Emojis to its own file
